+++ packages/salesforcedx-vscode-core/src/commands/forceSourceRetrieveManifest.ts
@@ -4,7 +4,6 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { LibraryCommandletExecutor } from '@salesforce/salesforcedx-utils-vscode/out/src';
 import {
   Command,
   SfdxCommandBuilder
@@ -13,19 +12,18 @@ import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/
 import { ComponentSet } from '@salesforce/source-deploy-retrieve';
 import { join } from 'path';
 import * as vscode from 'vscode';
-import { channelService, OUTPUT_CHANNEL } from '../channels';
+import { channelService } from '../channels';
 import {
   ConflictDetectionChecker,
   ConflictDetectionMessages
 } from '../commands/util/postconditionCheckers';
-import { workspaceContext } from '../context';
 import { nls } from '../messages';
 import { notificationService } from '../notifications';
 import { SfdxPackageDirectories } from '../sfdxProject';
 import { telemetryService } from '../telemetry';
 import { getRootWorkspacePath } from '../util';
+import { RetrieveExecutor } from './baseDeployRetrieve';
 import {
-  createRetrieveOutput,
   FilePathGatherer,
   SfdxCommandlet,
   SfdxCommandletExecutor,
@@ -46,37 +44,26 @@ export class ForceSourceRetrieveManifestExecutor extends SfdxCommandletExecutor<
   }
 }
 
-export class LibrarySourceRetrieveManifestExecutor extends LibraryCommandletExecutor<
+export class LibrarySourceRetrieveManifestExecutor extends RetrieveExecutor<
   string
 > {
   constructor() {
     super(
-      'Retrieve With Manifest (beta)',
-      'force_source_retrieve_with_manifest_beta',
-      OUTPUT_CHANNEL
+      nls.localize('force_source_retrieve_text'),
+      'force_source_retrieve_with_manifest_beta'
     );
   }
 
-  public async run(response: ContinueResponse<string>): Promise<boolean> {
+  protected async getComponents(
+    response: ContinueResponse<string>
+  ): Promise<ComponentSet> {
     const packageDirs = await SfdxPackageDirectories.getPackageDirectoryPaths();
-    const defaultOutput = join(
-      getRootWorkspacePath(),
-      (await SfdxPackageDirectories.getDefaultPackageDir()) ?? ''
-    );
-    const components = await ComponentSet.fromManifestFile(response.data, {
+    return ComponentSet.fromManifestFile(response.data, {
       resolve: packageDirs.map(relativeDir =>
         join(getRootWorkspacePath(), relativeDir)
       ),
       literalWildcard: true
     });
-    const connection = await workspaceContext.getConnection();
-    const result = await components.retrieve(connection, defaultOutput, {
-      merge: true
-    });
-
-    channelService.appendLine(createRetrieveOutput(result, packageDirs));
-
-    return result.success;
   }
 }