+++ packages/salesforcedx-vscode-lwc/test/vscode-integration/testSupport/mocks/index.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { shared as lspCommon } from '@salesforce/lightning-lsp-common';
-import { TestRunner as UtilsTestRunner } from '@salesforce/salesforcedx-utils-vscode/out/src/cli/';
+import * as pathUtils from '@salesforce/salesforcedx-utils-vscode/out/src/helpers';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as which from 'which';
@@ -125,7 +125,7 @@ export function unmockActiveTextEditorUri() {
  * Mock test result watcher's get temp folder and watch test results methods
  */
 export function mockTestResultWatcher() {
-  getTempFolderStub = stub(UtilsTestRunner.prototype, 'getTempFolder');
+  getTempFolderStub = stub(pathUtils, 'getTestResultsFolder');
   getTempFolderStub.callsFake((vscodePath: string, testType: string) => {
     return path.join(vscodePath, '.sfdx', 'tools', 'testresults', testType);
   });